+++ handlers/bill_handlers/get_bills_by_time.go
@@ -27,10 +27,10 @@ type TimeData struct {
 }
 
 type TimeRecord struct {
-	Receipt      int    `json:"receipt"`
-	Disbursement int    `json:"disbursement"`
-	Type         string `json:"type"`
-	Time         string `json:"time"`
+	Receipt      float64 `json:"receipt"`
+	Disbursement float64 `json:"disbursement"`
+	Type         string  `json:"type"`
+	Time         string  `json:"time"`
 }
 
 func (res *TimeResponse) setTimeResponse(code int, msg string, records []TypeRecord) {
